Fix required errors on IndicatorForm
